requirements.txt
vimgolf==0.5.1
agg-python-bindings==0.1.4
ptyprocess==0.7.0
pillow
pydantic
parse==1.20.2
# gym==0.26.2 # optionally use gymnasium==1.2.0 (actively maintained at: https://github.com/Farama-Foundation/Gymnasium)